Fixed all warnings on Mac/Clang and Linux/GCC

+++ libswirl/archive/7zArchive.cpp
@@ -48,7 +48,7 @@ ArchiveFile* SzArchive::OpenFile(const char* name)
 		if (isDir)
 			continue;
 
-		int len = SzArEx_GetFileNameUtf16(&szarchive, i, fname);
+		size_t len = SzArEx_GetFileNameUtf16(&szarchive, i, fname);
 		char szname[512];
 		int j = 0;
 		for (; j < len && j < sizeof(szname) - 1; j++)
@@ -63,7 +63,7 @@ ArchiveFile* SzArchive::OpenFile(const char* name)
 		if (res != SZ_OK)
 			return NULL;
 
-		return new SzArchiveFile(out_buffer, offset, (u32)out_size_processed);
+		return new SzArchiveFile(out_buffer, (u32)offset, (u32)out_size_processed);
 	}
 	return NULL;
 }
