+++ libswirl/deps/libzip/zip_source_zip.c
@@ -120,7 +120,7 @@ read_zip(void *state, void *data, size_t len, enum zip_source_cmd cmd)
 {
     struct read_zip *z;
     char b[8192], *buf;
-    int i, n;
+    ssize_t i, n;
 
     z = (struct read_zip *)state;
     buf = (char *)data;
